+++ src/Exception/AuthorizationException.php
@@ -8,4 +8,4 @@ namespace Xakki\FileUploader\Exception;
  * The current user is not allowed to manage (delete/restore) a file.
  * Bindings map this to HTTP 403.
  */
-class AuthorizationException extends UploadException {}
+class AuthorizationException extends CodedUploadException {}